requirements.txt
fastapi==0.115.6
Flask==3.1.0
pydantic==2.10.5
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
regex==2024.11.6
Requests==2.32.3
spacy==3.8.3
twikit==2.2.1
uvicorn==0.34.0
